Preliminary refactoring complete.
Made many changes, including the following:
- Added folder structure.
- Removed lots of redundant code.
- Removed lots of unused usings.
- Unified code style.
- Extracted Random into a static wrapper so as to always call next on the same seed.
Split some stuff apart, extended PieceType.

What really needs to be done:
- Refactor Application.cs
- Refactor Grid.cs
- Separate UI from Logic in the following files:
-- LevelText.cs
-- RealTimeText.cs
-- ScoreText.cs
-- StatsTextBlock.cs
