+++ TetrisDotnet/Code/Game/GameState.cs
@@ -1,0 +1,8 @@
+﻿namespace TetrisDotnet.Code.Game
+{
+	public enum GameState
+	{
+		Playing,
+		Pause
+	}
+}
