+++ TetrisDotnet/Code/Game/Hold.cs
@@ -1,0 +1,14 @@
+﻿namespace TetrisDotnet.Code.Game
+{
+	class Hold
+	{
+		public PieceType currentPiece { get; set; }
+
+		public bool canSwap { get; set; }
+
+		public Hold(PieceType piece = PieceType.Empty)
+		{
+			currentPiece = piece;
+		}
+	}
+}
