+++ TetrisDotnet/Code/Program.cs
@@ -1,0 +1,12 @@
+﻿namespace TetrisDotnet.Code
+{
+	internal static class Program
+	{
+		public static void Main()
+		{
+			Application app = new Application();
+
+			app.Run();
+		}
+	}
+}
