--- TetrisDotnet/Hold.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TetrisDotnet;
-
-namespace Tetris
-{
-    class Hold
-    {
-        public PieceType currentPiece { get; set; }
-
-        public bool canSwap { get; set; }
-
-        public Hold(PieceType piece = PieceType.Empty)
-        {
-            currentPiece = piece;
-        }
-    }
-}
