+++ TetrisDotnet/IconGenerator.cs
@@ -1,11 +1,5 @@
 ﻿using SFML.Graphics;
-using System;
-using System.Collections.Generic;
 //using System.Drawing;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Tetris
 {
