--- TetrisDotnet/PieceType.cs
@@ -1,16 +1,0 @@
-namespace TetrisDotnet
-{
-    public enum PieceType
-    {
-        I = 0,
-        O = 1,
-        T = 2,
-        S = 3,
-        Z = 4,
-        J = 5,
-        L = 6,
-        Empty = 7,
-        Ghost = 8,
-        Dead
-    }
-}
