--- TetrisDotnet/Program.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TetrisDotnet;
-
-namespace Tetris
-{
-    class Program
-    {
-        public static void Main(string[] args)
-        {
-            Application app = new Application();
-
-            app.Run();
-        }
-    }
-}
